desktopApp/requirements.txt
requests
requests-oauthlib
python-twitter
PyQt5
pycryptodome
mysql-connector-python
csv
